requirements.txt
streamlit>=1.40,<2.0
chromadb>=1.5,<2.0
pandas>=2.0,<3.0
langgraph>=1.0,<2.0
langchain-core>=0.3,<0.4
langchain-openai>=1.0,<2.0
PyYAML>=6,<7
openai>=1.0,<2.0
SpeechRecognition>=3.16,<4.0
pocketsphinx>=5.0,<6.0
